# tools/ftd-level-editor/openapi.json
{
  "components": {
    "schemas": {
...
      "EditorStatus": {
        "additionalProperties": false,
        "properties": {
          "providerMode": {
            "title": "Providermode",
            "type": "string"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "stores": {
            "items": {
              "type": "string"
            },
            "title": "Stores",
            "type": "array"
          },
          "workerMode": {
            "title": "Workermode",
            "type": "string"
          }
        },
        "required": [
          "service",
          "providerMode",
          "workerMode",
          "stores"
        ],
        "title": "EditorStatus",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "LaunchCredential": {
        "in": "header",
        "name": "X-FTD-Launch-Credential",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Find the Dog Level Editor",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/status": {
      "get": {
        "operationId": "getEditorStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "LaunchCredential": []
          }
        ],
        "summary": "Status",
        "x-ftd-authorization": "launch-credential",
        "x-ftd-cost": "none",
        "x-ftd-side-effects": "none"
      }
    },
...
  }
}